There is a race-condition where pointAnnotationManager is called before it's been set.
This is due to pointAnnotationManager requiring async to load, and setStyle callback being
called before it loaded. The simpliest fix was just to move the setStyle callback code into
the _onMapCreated callback as it was possible to set everything up in there.

There is still a subtle race if a user presses car_detailed FloatingActionButton button. I
think the correct fix is to disable that button until loading is done.