Fixed race conditions in AnimatedRoute (#246)
* Changed the file to consistently use ' instead of ". The file previously used both, and ' is preferred in the dart style guide.

* Listed the parameter's types in the constructor so they wouldn't be considered "dynamic", and improves IDE/linting.

This could be avoided by using `super.id` but that's not supported in Dart pre 2.17.0. I also didn't notice until finishing that these files are generated, but its unclear what
generates them.

* There is a race-condition where pointAnnotationManager is called before it's been set.

This is due to pointAnnotationManager requiring async to load, and setStyle callback being
called before it loaded. The simpliest fix was just to move the setStyle callback code into
the _onMapCreated callback as it was possible to set everything up in there.

There is still a subtle race if a user presses car_detailed FloatingActionButton button. I
think the correct fix is to disable that button until loading is done.

* Fixed another race condition, where if a PointAnnotation was clicked multiple times, the StyleLayer would be removed, but the Tween animation would still try and update it.

I also simplified the code a little, as the style was inconsient with above, and had a branch that could never be visited.

---------

Co-authored-by: Roman Laitarenko <roman.laitarenko@mapbox.com>